# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Parali Se Paisa Action Platform API",
    "description": "Offline-first prototype API. Demo records remain marked synthetic; this is not an operational government service.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Login",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "authentication"
        ],
        "summary": "Me",
        "operationId": "me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserView"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/system-status": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "System Status",
        "operationId": "system_status_api_v1_system_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/product": {
      "get": {
        "tags": [
          "product"
        ],
        "summary": "Product",
        "operationId": "product_api_v1_product_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductSnapshot"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/fields": {
      "get": {
        "tags": [
          "fields"
        ],
        "summary": "List Fields",
        "operationId": "list_fields_api_v1_fields_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/balers": {
      "get": {
        "tags": [
          "balers"
        ],
        "summary": "List Balers",
        "operationId": "list_balers_api_v1_balers_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "balers"
        ],
        "summary": "Create Baler",
        "operationId": "create_baler_api_v1_balers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BalerCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/buyers": {
      "get": {
        "tags": [
          "buyers"
        ],
        "summary": "List Buyers",
        "operationId": "list_buyers_api_v1_buyers_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "buyers"
        ],
        "summary": "Create Buyer",
        "operationId": "create_buyer_api_v1_buyers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuyerCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/jobs": {
      "get": {
        "tags": [
          "dispatch",
          "collection"
        ],
        "summary": "List Jobs",
        "operationId": "list_jobs_api_v1_jobs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/certificates": {
      "get": {
        "tags": [
          "certificates"
        ],
        "summary": "Certificates List",
        "operationId": "certificates_list_api_v1_certificates_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "BalerCreate": {
        "properties": {
          "baler_id": {
            "type": "string",
            "title": "Baler Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "operator_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operator Name"
          },
          "latitude": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude"
          },
          "current_location_lat": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 90.0,
                "minimum": -90.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Location Lat"
          },
          "current_location_lon": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 180.0,
                "minimum": -180.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Location Lon"
          },
          "status": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "BUSY",
              "OFFLINE",
              "MAINTENANCE"
            ],
            "title": "Status",
            "default": "AVAILABLE"
          },
          "working_start": {
            "type": "string",
            "title": "Working Start",
            "default": "08:00"
          },
          "working_end": {
            "type": "string",
            "title": "Working End",
            "default": "17:00"
          },
          "service_rate_acres_per_hour": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Service Rate Acres Per Hour"
          },
          "daily_capacity_acres": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Daily Capacity Acres"
          },
          "max_straw_capacity_tonnes": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Straw Capacity Tonnes"
          },
          "available_from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Available From"
          },
          "demo_or_real": {
            "type": "string",
            "enum": [
              "DEMO",
              "SYNTHETIC",
              "FIXTURE",
              "REAL"
            ],
            "title": "Demo Or Real",
            "default": "DEMO"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "API_UNVERIFIED_INPUT"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "baler_id",
          "name",
          "latitude",
          "longitude",
          "service_rate_acres_per_hour",
          "daily_capacity_acres"
        ],
        "title": "BalerCreate"
      },
...
      "BuyerCreate": {
        "properties": {
          "buyer_id": {
            "type": "string",
            "title": "Buyer Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "buyer_type": {
            "type": "string",
            "enum": [
              "BIOMASS_POWER",
              "CBG",
              "PELLET",
              "INDUSTRIAL_BOILER",
              "OTHER"
            ],
            "title": "Buyer Type",
            "default": "OTHER"
          },
          "latitude": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude"
          },
          "accepted_residue_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Accepted Residue Types"
          },
          "daily_demand_tonnes": {
            "type": "number",
            "minimum": 0.0,
            "title": "Daily Demand Tonnes"
          },
          "remaining_demand_tonnes": {
            "type": "number",
            "minimum": 0.0,
            "title": "Remaining Demand Tonnes"
          },
          "max_moisture_pct": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Moisture Pct"
          },
          "price_per_tonne": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Per Tonne"
          },
          "price_is_verified": {
            "type": "boolean",
            "title": "Price Is Verified",
            "default": false
          },
          "operating_start": {
            "type": "string",
            "title": "Operating Start",
            "default": "00:00"
          },
          "operating_end": {
            "type": "string",
            "title": "Operating End",
            "default": "23:59"
          },
          "status": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "UNAVAILABLE",
              "PAUSED"
            ],
            "title": "Status",
            "default": "AVAILABLE"
          },
          "demo_or_real": {
            "type": "string",
            "enum": [
              "DEMO",
              "SYNTHETIC",
              "FIXTURE",
              "REAL"
            ],
            "title": "Demo Or Real",
            "default": "DEMO"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "API_UNVERIFIED_INPUT"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "buyer_id",
          "name",
          "latitude",
          "longitude",
          "daily_demand_tonnes",
          "remaining_demand_tonnes"
        ],
        "title": "BuyerCreate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest"
      },
      "LoginResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In",
            "default": 1800
          },
          "user": {
            "$ref": "#/components/schemas/UserView"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "user"
        ],
        "title": "LoginResponse"
      },
...
      "ProductSnapshot": {
        "properties": {
          "fields": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Fields"
          },
          "balers": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Balers"
          },
          "buyers": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Buyers"
          },
          "jobs": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Jobs"
          },
          "runs": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Runs"
          },
          "verification": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Verification"
          },
          "certificates": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Certificates"
          },
          "allocations": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Allocations"
          },
          "pickup_requests": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Pickup Requests"
          },
          "stats": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Stats"
          },
          "generated_at": {
            "type": "string",
            "title": "Generated At"
          }
        },
        "type": "object",
        "required": [
          "fields",
          "balers",
          "buyers",
          "jobs",
          "runs",
          "verification",
          "certificates",
          "allocations",
          "pickup_requests",
          "stats",
          "generated_at"
        ],
        "title": "ProductSnapshot"
      },
      "UserView": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "field_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field Id"
          },
          "baler_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baler Id"
          },
          "buyer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Buyer Id"
          },
          "is_demo": {
            "type": "boolean",
            "title": "Is Demo"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "username",
          "role",
          "field_id",
          "baler_id",
          "buyer_id",
          "is_demo"
        ],
        "title": "UserView"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}